+++ packages/config/src/discovery/utils.test.ts
@@ -1,0 +1,16 @@
+import { expect } from 'earl'
+import { trimTrailingDots } from './utils'
+
+describe(trimTrailingDots.name, () => {
+  it('should remove trailing dots', () => {
+    const description = 'Some description...'
+    const trimmed = trimTrailingDots(description)
+    expect(trimmed).toEqual('Some description')
+  })
+
+  it('should not remove trailing dots if there are no dots', () => {
+    const description = 'Some description'
+    const trimmed = trimTrailingDots(description)
+    expect(trimmed).toEqual(description)
+  })
+})
